feat: refactor the code to modules and add search and sorting
