+++ .vscode/settings.json
@@ -1,0 +1,9 @@
+{
+    "files.associations": {
+        "ostream": "cpp",
+        "array": "cpp",
+        "sstream": "cpp",
+        "tuple": "cpp",
+        "stdexcept": "cpp"
+    }
+}
