+++ src/utils.h
@@ -1,0 +1,12 @@
+#ifndef UTILS_H
+#define UTILS_H
+
+#include <iostream>
+using namespace std;
+
+void clearScreen();
+int generateRandomOrderId();
+
+#include "utils.cpp"
+
+#endif
